% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bulk.R
\name{Bulk}
\alias{Bulk}
\title{GNAF bulk request}
\usage{
Bulk(df, boundingbox, fuzzy = FALSE)
}
\arguments{
\item{Addresses}{Full address string including street number (and or unit number. building name etc.), street name, suburb and postcode.
boundingbox a named list of numeric coordinates in WGS 84 projection marking the minimum bounding rectangle of the search area. The
list should be in the format: minLat, minLon, maxLat, maxLon.
fuzzy puts boundaries around string matching parameters I.e. MaxEdits, MinLength and prefixLength. Defaults to FALSE}
}
\description{
This function communicates with the GNAF geocoding API to return an official address.
}
\examples{
Bulk()
}
\keyword{Address}
\keyword{GNAF,}
\keyword{Geocode,}
\keyword{fuzzymatching,}
