% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSONquery.R
\name{JSONquery}
\alias{JSONquery}
\title{GNAF Geocode JSON query preparation}
\usage{
JSONquery(x = x, y = y, id = id, FULL_ADDRESS = FULL_ADDRESS, radius,
  convert = TRUE)
}
\arguments{
\item{x}{longitude coordinate column in dataframe.
y latitude coordinate column in dataframe
id the unique identifier columns to be used
FULL_ADDRESS the address string for input into the GNAF Geocode function. Full address string including street number (and or unit number. building name etc.), street name, suburb and postcode. NOTE: remove all "/"'s from string for better performance.
radius the search radius for the GNAF geocode function to use in meters around the input coordinate.
convert are the input coordinates projected in CRS UTMGRS80 zone 55? Defaults to FALSE}
}
\description{
This function prepares a set of coordinates and address strings to feed into the GNAF geocode function. the result is a list of addresses with their corresponding minimum bounding rectangle for a given radius around a coordinate (x,y).
}
\examples{
JSONquery()
}
\keyword{GNAF,}
\keyword{Geocode}
\keyword{JSON}
\keyword{boundingbox,}
\keyword{queary,}
