% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{Process}
\alias{Process}
\title{Process the JSON result from the GNAF Geocode function}
\usage{
Process(response)
}
\arguments{
\item{responses}{list of JSON responeses from the GNAF Geocode function.}
}
\description{
This function takes the resulting list of JSON responses from the GNAF Geocode function and returns a list of results.
}
\examples{
Process()

}
\keyword{GNAF}
\keyword{GNAF,}
\keyword{Geocode}
\keyword{JSON,}

