% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VICMAPtest.R
\name{VICMaps}
\alias{VICMaps}
\title{GNAF Geocode compare to VICMaps}
\usage{
VICMaps(spdf = VicMAPS)
}
\arguments{
\item{VicMAPs}{SpatialPointsDataFrame of the VicMAPS shapefile}
}
\description{
This function returns a sample of 1000 addresses from the VICmaps address shapefile and geocodes them using the GNAF API
}
\examples{
VICMaps()

}
\keyword{GNAF,}
\keyword{VICMaps}

